backend/ai_face_service/requirements.txt
flask
psycopg2-binary
numpy
mediapipe
opencv-python
onnxruntime   # important!
